refactor: remove unused rules and processors from plugin configuration; update versioning and import paths
