+++ packages/eslint-plugin-angular-template/src/index.ts
@@ -2,8 +2,6 @@ import recommended from './configs/recommended.js';
 
 const plugin = {
 	configs: { recommended },
-	rules: {},
-	processors: {},
 };
 
 export default plugin;
