+++ packages/eslint-plugin/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tylertech-eslint/eslint-plugin",
-  "version": "1.0.4",
+  "version": "1.0.9",
   "description": "Common TypeScript lint rules/preferences for Tyler Technologies. Compatible with ESLint v9+ (Flat Config only).",
   "type": "module",
   "exports": {
