+++ packages/eslint-plugin/src/index.ts
@@ -1,10 +1,7 @@
 import recommended from './configs/recommended.js';
-import { rules } from './rules/index.js';
 
 const plugin = {
 	configs: { recommended },
-	rules,
-	processors: {},
 };
 
 export default plugin;
