+++ packages/eslint-plugin/src/rules/index.ts
@@ -1,9 +1,11 @@
-import invalidRelativeImportPrefix, { RULE_NAME as invalidRelativeImportPrefixRuleName } from './invalid-relative-Import-prefix.js';
+import invalidRelativeImportPrefix, { RULE_NAME as invalidRelativeImportPrefixRuleName } from './invalid-relative-import-prefix.js';
 import requirePrivateUnderscore, { RULE_NAME as requirePrivateUnderscoreRuleName } from './require-private-underscore.js';
 import requirePrivateModifier, { RULE_NAME as requirePrivateModifierRuleName } from './require-private-modifier.js';
 
-export const rules = {
+export default {
+  rules: {
     [requirePrivateModifierRuleName]: requirePrivateModifier,
     [requirePrivateUnderscoreRuleName]: requirePrivateUnderscore,
     [invalidRelativeImportPrefixRuleName]: invalidRelativeImportPrefix
+  }
 };
